09-serverless/requirements.txt
onnxruntime
torch==2.8.0
torchvision
pillow